requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
boto3==1.23.10
botocore==1.26.10
colorama==0.4.4
decorator==5.1.1
distlib==0.3.4
Django==4.0.4
executing==0.8.3
filelock==3.7.0
ipython==8.3.0
jedi==0.18.1
jmespath==1.0.0
matplotlib-inline==0.1.3
PyMySQL==1.0.2
mysqlclient==2.1.0
numpy==1.22.4
opencv-python==4.5.5.64
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
prompt-toolkit==3.0.29
pure-eval==0.2.2
Pygments==2.12.0
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.2.1.post0
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5

joblib == 1.1.0 
scikit-learn == 1.1.1 
sklearn == 0.0 
threadpoolctl == 3.1.0

# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1  # Google Colab version
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow